cleanup

+++ board/safety/safety_hyundai_common.h
@@ -84,7 +84,7 @@ void hyundai_common_cruise_buttons_check(const int cruise_button, const int main
   }
 }
 
-uint16_t hyundai_common_canfd_compute_checksum(CANPacket_t *to_push) {
+uint32_t hyundai_common_canfd_compute_checksum(CANPacket_t *to_push) {
   int len = GET_LEN(to_push);
   uint32_t address = GET_ADDR(to_push);
